router-config: Bump version info
The Edgerouter will try to "migrate" configuration to the latest
possible version when it is loaded. That has been failing recently
with the following error:

> Jun  4 20:52:25 router0 config-migrate: Config file /config/config.pyinfra is not writable. Migration aborted.

Changing the file to mode 0660 and loading it does the trick.

Rather than let the router migrate configuration when it wants, I will
migrate versions manually so that /config/config.pyinfra matches up
with the template in this repository. If there were a significant
syntax change and the configurations did not line up, that would be
quite confusing, indeed.